chore: Add `select-item-custom` example
