--- packages/ariakit/src/select/__examples__/select-dialog-form/style.css
@@ -1,130 +1,0 @@
-@import url("../select/style.css");
-@import url("../../../form/__examples__/form/style.css");
-
-.wrapper {
-  @apply my-[800px];
-}
-
-.photo {
-  @apply w-12 h-12 rounded-full;
-}
-
-.select {
-  width: 320px;
-  @apply justify-start h-auto gap-3 p-4;
-}
-
-.popover {
-  width: 320px;
-  transform: translateY(-16px);
-  opacity: 0;
-  transform-origin: top center;
-  transition: transform 200ms ease-in-out, opacity 200ms ease-in-out;
-}
-
-.popover[data-enter] {
-  transform: translateY(0);
-  opacity: 1;
-}
-
-/* .popover > * {
-  min-height: 60px;
-  @apply gap-3;
-} */
-
-.select-item {
-  min-height: 60px;
-  @apply
-    flex
-    gap-3
-    items-center
-    scroll-m-[0.1px]
-    outline-none
-    cursor-default
-    rounded
-    p-2
-    aria-disabled:opacity-50;
-}
-
-.popover:not([aria-activedescendant]) .select-item:hover,
-.select-item[data-active-item] {
-  @apply
-    bg-primary-2
-    text-primary-2
-    dark:bg-primary-2-dark
-    dark:text-primary-2-dark;
-}
-
-.value {
-  @apply flex flex-col items-start flex-1;
-}
-
-.email {
-  @apply opacity-80;
-}
-
-[data-backdrop] {
-  background-color: rgb(0 0 0 / 10%);
-}
-
-.dialog {
-  max-width: min(calc(100vw - 16px), 320px);
-  max-height: calc(100vh - 16px);
-  opacity: 0;
-  transform: scale(0);
-  transform-origin: center 0;
-  transition: opacity 200ms ease-in-out, transform 200ms ease-in-out;
-  @apply
-    w-80
-    h-fit
-    fixed
-    inset-0
-    m-auto
-    z-50
-    flex
-    flex-col
-    gap-4
-    rounded-lg
-    border
-    border-solid
-    border-canvas-4
-    bg-canvas-4
-    p-4
-    text-canvas-4
-    drop-shadow-lg
-    focus-visible:ariakit-outline
-    dark:border-canvas-4-dark
-    dark:bg-canvas-4-dark
-    dark:text-canvas-4-dark
-    dark:drop-shadow-lg-dark;
-}
-
-.dialog[data-enter] {
-  transform: scaleY(1);
-  opacity: 1;
-}
-
-.dialog header {
-  @apply flex items-center justify-between;
-}
-
-.dismiss {
-  @apply
-    p-2
-    rounded-lg
-    bg-alpha-2
-    text-alpha-2
-    hover:bg-alpha-2-hover
-    dark:bg-alpha-2-dark
-    dark:text-alpha-2-dark
-    dark:hover:bg-alpha-2-dark-hover
-    focus-visible:ariakit-outline;
-}
-
-.field input {
-  @apply w-full;
-}
-
-.heading {
-  @apply m-0 text-xl font-medium;
-}
